Added download callbacks and progress bar.
Added handlers (callback subroutines) for when a header is downloaded
(response_header), part of the content is downloaded (response_data), and when
the content has finished downloading (response_done).  Each of these callback
methods update a progress bar to indicate how far the download has progressed.

The progress bar is recreated for each file that is downloaded.  The script
should be fault-tolerant enough to avoid errors if the progress bar could not
be created.

This code needs to be cleaned up (and split into additional subroutines).